Keep as little documentation as possible in the README.md
The less documentation is available in the README.md, the more likely people
are to actually read the extensive documentation.

There's no straightforward way to do reverse SSH tunnels and the usage is
available via the rssht command line anyway, so there's is no point to
keep the usage instructions in the README.md.