Use a host alias to support multiple tunnels to the same host
The alias is of the form rssht_$host_$port.

This change requires disabling the check of the host IP. While
this doesn't seem to be a big deal security-wise given the server's
identity is checked using its public key anyway, this is still
removing a security check. To be reworked later for stronger security.

This has zero security impact if using IP to connect instead of DNS.