Use a key alias when connecting over an HTTP tunnel
Since the host is always localhost and the port is random, one
would have to add the host / port to ~/.ssh/known_hosts everytime
or even have conflicts because the same port is reused for tunnels
to different hosts. The alias using the actual hostname and port
(rather than the tunnel's) solves this.

There's no reason to worry about CheckHostIP there, since the IP
is always the loopback address (127.0.0.1 or ::1).